+++ core/src/test/java/feign/FeignTest.java
@@ -535,7 +535,7 @@ public class FeignTest {
           public Object decode(Response response, Type type) throws IOException {
             String string = super.decode(response, type).toString();
             if ("retry!".equals(string)) {
-              throw new RetryableException(response.status(), string, HttpMethod.POST, null,
+              throw new RetryableException(response.status(), string, HttpMethod.POST, (Long) null,
                   response.request());
             }
             return string;
@@ -616,7 +616,7 @@ public class FeignTest {
           @Override
           public Exception decode(String methodKey, Response response) {
             return new RetryableException(response.status(), "play it again sam!", HttpMethod.POST,
-                null, response.request());
+                (Long) null, response.request());
           }
         }).target(TestInterface.class, "http://localhost:" + server.getPort());
 
@@ -641,7 +641,7 @@ public class FeignTest {
           @Override
           public Exception decode(String methodKey, Response response) {
             return new RetryableException(response.status(), "play it again sam!", HttpMethod.POST,
-                new TestInterfaceException(message), null, response.request());
+                new TestInterfaceException(message), (Long) null, response.request());
           }
         }).target(TestInterface.class, "http://localhost:" + server.getPort());
 
@@ -663,7 +663,7 @@ public class FeignTest {
         .errorDecoder(new ErrorDecoder() {
           @Override
           public Exception decode(String methodKey, Response response) {
-            return new RetryableException(response.status(), message, HttpMethod.POST, null,
+            return new RetryableException(response.status(), message, HttpMethod.POST, (Long) null,
                 response.request());
           }
         }).target(TestInterface.class, "http://localhost:" + server.getPort());
